% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_disciplina.R
\name{conta_votacoes_sem_consenso}
\alias{conta_votacoes_sem_consenso}
\title{Conta votações sem consenso}
\usage{
conta_votacoes_sem_consenso(votos)
}
\arguments{
\item{votos}{Dataframe de votos. Devem ter pelo menos 3 colunas: id_votacao, casa e voto.}
}
\value{
Dataframe indicando a quantidade de votações sem consenso por casa
}
\description{
Conta quantas votações ocorerram em quais não há consenso
}
\examples{
conta_votacoes_sem_consenso(votos)
}
