% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_orientacoes.R
\name{define_orientacao_governo}
\alias{define_orientacao_governo}
\title{Define a orientação do Governo}
\usage{
define_orientacao_governo(votos)
}
\arguments{
\item{votos}{Dataframe contendo dados de votos}
}
\value{
Dataframe contendo informações de orientações do Governo
}
\description{
Recupera informação das orientações do Governo com base nos votos do Líder
}
\examples{
orientacao <- define_orientacao_governo(votos)
}
