% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_cargos}
\alias{.check_cargos}
\title{Checa se cargos da mesa se encaixam na lista adotada como padrão}
\usage{
.check_cargos(df)
}
\arguments{
\item{df}{Dataframe com pelo menos uma coluna chamada cargo}
}
\value{
Dataframe com mesmo conteúdo que o passado como parâmetro.
}
\description{
avalia se cargos presentes como coluna em um dataframe se encaixam na lista adotada como padrão.
Lança um erro caso seja encontrado um cargo que não se encaixe na lista de cargos possíveis.
}
