% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_comissoes.R
\name{.enumera_cargo_comissao}
\alias{.enumera_cargo_comissao}
\title{Atribui um peso ao cargo do parlamentar na Comissão}
\usage{
.enumera_cargo_comissao(cargo, situacao)
}
\arguments{
\item{cargo}{Cargo para padronização}

\item{situacao}{Situação do parlamentar na Comissão}
}
\value{
Valor atribuído ao cargo
}
\description{
Classifica cargos da Comissão de acordo com o nível do cargo
}
\examples{
.enumera_cargo_comissao("Titular")
}
