% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_liderancas_camara.R
\name{.fetch_liderancas_camara}
\alias{.fetch_liderancas_camara}
\title{Recupera as informações sobre lideranças de blocos e partidos}
\usage{
.fetch_liderancas_camara()
}
\value{
Dataframe contendo lideranças dos blocos e partidos na Câmara
}
\description{
Retorna um dataframe contendo os líderes, vice-líderes e representantes dos blocos e partidos na Câmara
}
\examples{
liderancas <- .fetch_liderancas_camara()
}
