% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_liderancas_senado.R
\name{.fetch_liderancas_senado}
\alias{.fetch_liderancas_senado}
\title{Recupera as informações sobre lideranças de blocos e partidos no Senado}
\usage{
.fetch_liderancas_senado()
}
\value{
Dataframe contendo lideranças dos blocos e partidos no Senado Federal
}
\description{
Retorna um dataframe contendo os líderes, vice-líderes e representantes dos blocos e partidos
no Senado Federal
}
\examples{
liderancas <- .fetch_liderancas_senado()
}
