% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_comissoes_camara.R
\name{.fetch_membros_comissao_camara_by_page}
\alias{.fetch_membros_comissao_camara_by_page}
\title{Recupera informações dos membros de uma Comissão em uma data de inicio específica}
\usage{
.fetch_membros_comissao_camara_by_page(
  page = 1,
  url,
  sigla_comissao,
  max_tentativas = 10
)
}
\arguments{
\item{page}{Página a ser requisitada}

\item{url}{Url da requisição}

\item{max_tentativas}{Número máximo de tentativas}
}
\value{
Dataframe com informações dos membros da Comissão para uma data de início
}
\description{
A partir de um id, retorna os membros daquela comissão a partir da data de início especificada
}
