% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_comissoes_camara.R
\name{.fetch_membros_comissao_camara_with_backoff}
\alias{.fetch_membros_comissao_camara_with_backoff}
\title{Recupera informações dos membros de uma Comissão em uma data de inicio específica}
\usage{
.fetch_membros_comissao_camara_with_backoff(
  orgao_id,
  sigla_comissao,
  data_inicio = "2019-12-22",
  max_tentativas = 10
)
}
\arguments{
\item{orgao_id}{Id da Comissão}

\item{sigla_comissao}{Sigla da Comissão}

\item{data_inicio}{Data inicial de interesse (formato AAAA-MM-DD)}

\item{max_tentativas}{Número máximo de tentativas}
}
\value{
Dataframe com informações dos membros da Comissão para uma data de início
}
\description{
A partir de um id, retorna os membros daquela comissão a partir da data de início especificada
}
