% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_comissoes.R
\name{.padroniza_cargo_comissao}
\alias{.padroniza_cargo_comissao}
\title{Padroniza nomenclatura do cargo de um parlamentar numa Comissão}
\usage{
.padroniza_cargo_comissao(cargo)
}
\arguments{
\item{cargo}{Cargo para padronização}
}
\value{
String com Nome Padronizado
}
\description{
Padroniza nomenclatura do cargo de um parlamentar numa Comissão
}
\examples{
.padroniza_cargo_comissao("Titular")
}
