% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_parlamentares_por_legislatura.R
\name{.processa_dados_senadores}
\alias{.processa_dados_senadores}
\title{Processa dados de senadores}
\usage{
.processa_dados_senadores(legislaturas_list = c(55, 56))
}
\arguments{
\item{legislaturas_list}{Lista com ids das legislaturas. Default = 55 e 56.}
}
\value{
Dataframe contendo informações sobre os senadores
}
\description{
Processa informações sobre os senadores das legislaturas 55 e 56
}
\examples{
.processa_dados_senadores()
}
