% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_governismo.R
\name{.processa_quantidade_votos}
\alias{.processa_quantidade_votos}
\title{Processa a quantidade de votos válidos dos parlamentares}
\usage{
.processa_quantidade_votos(votos)
}
\arguments{
\item{votos}{Dataframe de votos
Os votos devem ter pelo menos 2 colunas: id_votacao e voto.}
}
\value{
Dataframe com o id do parlamentar e quantas vezes ele votou
}
\description{
Processa o número total de votações para cada parlamentar
}
\examples{
.processa_quantidade_votos(votos)
}
