% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enumera_voto}
\alias{enumera_voto}
\title{Enumera votações}
\usage{
enumera_voto(df)
}
\arguments{
\item{df}{Dataframe com a coluna voto}
}
\value{
Dataframe com coluna voto enumerada
}
\description{
Recebe um dataframe com coluna voto e enumera o valor para um número
}
\examples{
enumera_votacoes(df)
}
