% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_proposicoes_votadas_senado.R
\name{extract_number_from_regex}
\alias{extract_number_from_regex}
\title{Extrai somente os números de um texto extraído de regex}
\usage{
extract_number_from_regex(text, text_regex)
}
\arguments{
\item{text}{Texto onde o regex será aplicado}

\item{text_regex}{Expressão regular onde o texto será extraído para depois serem retornados apenas os números}
}
\value{
Números existentes em um texto extraído a partir de uma expressão regular
}
\description{
A partir de uma expressão regular, extrai o texto desse regex e depois retorna apenas os números existentes
}
\examples{
extract_number_from_regex("p_cod_materia_i=[\\\\d]*&")
}
