% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_text_from_node}
\alias{extract_text_from_node}
\title{Extrai o texto de um nó a partir do xpath}
\usage{
extract_text_from_node(node, xpath)
}
\arguments{
\item{node}{Nó do XML}

\item{xpath}{Xpath onde o texto está}
}
\value{
Texto extraído
}
\description{
Recebe um nó XML e um xpath e retorna o texto do conteúdo
}
