% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_text_from_pdf_url}
\alias{extract_text_from_pdf_url}
\title{Recebe uma URL para um pdf e retorna o texto raspado do conteúdo}
\usage{
extract_text_from_pdf_url(url)
}
\arguments{
\item{url}{URL contendo o pdf}
}
\value{
Texto do conteúdo do pdf
}
\description{
Recebe uma URL para um pdf e retorna o texto raspado do seu conteúdo
}
\examples{
extract_text_from_pdf_url("https://www.camara.leg.br/proposicoesWeb/prop_mostrarintegra?codteor=1709372")
}
