% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_proposicoes_info_senado.R
\name{fetch_all_proposicoes}
\alias{fetch_all_proposicoes}
\title{Recupera e processa dados de um conjunto de proposições}
\usage{
fetch_all_proposicoes(ids)
}
\arguments{
\item{ids}{lista de ids de proposições}
}
\value{
Dataframe com os dados de proposições
}
\description{
A partir de uma lista de ids, retorna os dados das respectivas proposições
}
\examples{
proposicoes <- fetch_all_proposicoes(ids)
}
