% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_proposicoes_info_senado.R
\name{fetch_all_proposicoes_votadas_em_intervalo_senado}
\alias{fetch_all_proposicoes_votadas_em_intervalo_senado}
\title{Recupera e processa dados de um conjunto de proposições que tiveram votações nominais em plenário em um intervalo}
\usage{
fetch_all_proposicoes_votadas_em_intervalo_senado(
  initial_date = "01/02/2019",
  end_date = format(Sys.Date(), "\%d/\%m/\%Y")
)
}
\arguments{
\item{initial_date}{Data inicial (formato "dd/mm/yyyy")}

\item{end_date}{Data final (formato "dd/mm/yyyy")}
}
\value{
Dataframe com os dados de proposições do intervalo
}
\description{
Retorna os dados das proposições que tiveram votações nominais em plenário em um intervalo
}
\examples{
proposicoes <- fetch_all_proposicoes_votadas_em_intervalo_senado()
}
