% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_cargos_mesa_camara.R
\name{fetch_cargos_mesa_camara}
\alias{fetch_cargos_mesa_camara}
\title{Captura informações de deputados integrantes da Mesa na Câmara dos deputados}
\usage{
fetch_cargos_mesa_camara(legislatura = 56, atual_cargo = TRUE)
}
\arguments{
\item{legislatura}{Número da legislatura para recuperação dos dados da mesa.}

\item{atual_cargo}{Flag que indica se deseja obter os deputados atualmente ocupando cargos na mesa. 
Se TRUE o parâmetro legislatura será ignorado e a legislatura 56 será usada.}
}
\value{
Dataframe contendo deputados com cargos na Mesa da Câmara.
}
\description{
Com base na API da Câmara dos Deputados captura informações de deputados integrantes da Mesa
}
\examples{
deputados_mesa <- fetch_cargos_mesa_camara()
}
