% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_cargos_mesa_senado.R
\name{fetch_cargos_mesa_senado}
\alias{fetch_cargos_mesa_senado}
\title{Captura informações de senadores atualmente integrantes da Mesa no Senado Federal}
\usage{
fetch_cargos_mesa_senado()
}
\value{
Dataframe contendo senadores com cargos na Mesa do Senado.
}
\description{
Com base na API da Senado captura informações de senadores integrantes da Mesa
}
\examples{
senadores_mesa <- fetch_cargos_mesa_senado()
}
