% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_comissoes.R, R/fetcher_comissoes.R
\name{fetch_comissao_info}
\alias{fetch_comissao_info}
\title{Recupera informações da Comissão}
\usage{
fetch_comissao_info(sigla, casa)

fetch_comissao_info(sigla, casa)
}
\arguments{
\item{sigla}{Sigla da Comissão}
}
\value{
Dataframe com informações da Comissão

Dataframe com informações da Comissão
}
\description{
Utiliza o rcongresso para recuperar informações sobre uma Comissão específica

Utiliza o rcongresso para recuperar informações sobre uma Comissão específica
}
\examples{
fetch_comissao_info("CCJC", "camara")
fetch_comissao_info("CCJC", "camara")
}
