% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_comissoes_camara.R
\name{fetch_comissao_info_camara}
\alias{fetch_comissao_info_camara}
\title{Recupera informações da Comissão na Câmara dos Deputados}
\usage{
fetch_comissao_info_camara(sigla_comissao)
}
\arguments{
\item{sigla_comissao}{Sigla da Comissão}
}
\value{
Dataframe com informações da Comissão
}
\description{
Utiliza o rcongresso para recuperar informações sobre uma Comissão específica na câmara dos deputados
}
\examples{
fetch_comissao_info_camara("CCJC")
}
