% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_comissoes.R,
%   R/fetcher_comissoes_senado.R
\name{fetch_comissao_info_senado}
\alias{fetch_comissao_info_senado}
\title{Recupera informações da Comissão no Senado Federal}
\usage{
fetch_comissao_info_senado(sigla)

fetch_comissao_info_senado(sigla)
}
\arguments{
\item{sigla}{Sigla da Comissão}
}
\value{
Dataframe com informações da Comissão

Dataframe com informações da Comissão
}
\description{
Utiliza o rcongresso para recuperar informações sobre uma Comissão específica no Senado

Utiliza o rcongresso para recuperar informações sobre uma Comissão específica no Senado
}
\examples{
fetch_comissao_info_senado("CAE")
fetch_comissao_info_senado("CAE")
}
