% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_comissoes.R,
%   R/fetcher_comissoes_camara.R
\name{fetch_comissoes_composicao_camara}
\alias{fetch_comissoes_composicao_camara}
\title{Recupera informações de composição das comissões usando o LeggoR a partir de uma data de inicio}
\usage{
fetch_comissoes_composicao_camara(data_inicio = "2019-12-22")

fetch_comissoes_composicao_camara(data_inicio = "2019-12-22")
}
\arguments{
\item{data_inicio}{Data inicial (formato AAAA-MM-DD)}
}
\value{
Dataframe com parlamentares membros da comissão e seus respectivos cargos

Dataframe com parlamentares membros da comissão e seus respectivos cargos
}
\description{
Utiliza o LeggoR para recuperar informações sobre a composição dos membros das composições do Congresso a partir de uma data

Utiliza o LeggoR para recuperar informações sobre a composição dos membros das composições do Congresso a partir de uma data
}
\examples{
fetch_comissoes_composicao_camara()
fetch_comissoes_composicao_camara()
}
