% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_deputados_info.R
\name{fetch_deputado}
\alias{fetch_deputado}
\title{Baixa nome civil dos deputados}
\usage{
fetch_deputado(id_deputado)
}
\arguments{
\item{id_votacao}{id do deputado}
}
\value{
Dataframe informações de id e nome civil.
}
\description{
Baixa nome civil dos deputados pelo id do parlamentar na Câmara.
}
\examples{
deputado <- fetch_deputado(73874)
}
