% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_deputados_info.R
\name{fetch_deputado_with_backoff}
\alias{fetch_deputado_with_backoff}
\title{Baixa dados dos deputados utilizando backoff exponencial}
\usage{
fetch_deputado_with_backoff(id_deputado, max_attempts = 10)
}
\arguments{
\item{id_votacao}{id do deputado}
}
\value{
Dataframe informações de id e nome civil.
}
\description{
Baixa nome civil dos deputados pelo id do parlamentar na Câmara utilizando backoff exponencial
com N tentativas (10 por padrão)
}
\examples{
deputado <- fetch_deputado_with_backoff(73874)
}
