% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_deputados_info.R
\name{fetch_deputados_legislatura}
\alias{fetch_deputados_legislatura}
\title{Importa dados de todos os deputados de uma legislatura específica}
\usage{
fetch_deputados_legislatura(legislatura = 56)
}
\value{
Dataframe contendo informações dos deputados: id, nome civil e cpf
}
\description{
Importa os dados de todos os deputados federais de uma legislatura específica
}
\examples{
deputados <- fetch_deputados_legislatura(56)
}
