% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_proposicoes_info_camara.R
\name{fetch_info_proposicao_camara}
\alias{fetch_info_proposicao_camara}
\title{Recupera informações de uma proposição}
\usage{
fetch_info_proposicao_camara(id_prop)
}
\arguments{
\item{id_prop}{ID de uma proposição}
}
\value{
Dataframe contendo informações de uma proposição
}
\description{
A partir do id, recupera dados de uma proposição na Câmara dos Deputados,
como nome, data_apresentacao, ementa, autor, indexacao, tema e uri_tramitacao
}
\examples{
proposicao <- fetch_info_proposicao_camara(2193540)
}
