% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_proposicoes_info_senado.R
\name{fetch_info_proposicao_senado}
\alias{fetch_info_proposicao_senado}
\title{Recupera e processa dados de uma proposição}
\usage{
fetch_info_proposicao_senado(id_proposicao)
}
\arguments{
\item{id_proposicao}{ID da proposição}
}
\value{
Dataframe com os dados de proposições
}
\description{
A partir de um id, retorna os dados de uma proposição
}
\examples{
proposicoes <- fetch_info_proposicao_senado(id_proposicao)
}
