% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_mandatos_camara.R
\name{fetch_mandatos_camara}
\alias{fetch_mandatos_camara}
\title{Extrai informações sobre o mandato de um deputado}
\usage{
fetch_mandatos_camara(id_deputado)
}
\arguments{
\item{id_deputado}{id do deputado}
}
\value{
Dataframe contendo informações de legislatura,
data de inicio, data de fim, situacao, código e descrição da causa do fim do exercício.
}
\description{
Recebe o id de um deputado e retorna um dataframe contendo informações do mandato, como legislatura,
data de inicio, data de fim, situacao, código e descrição da causa do fim do exercício.
}
\examples{
fetch_mandatos_camara(141428)
}
