% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_orientacoes_camara.R
\name{fetch_orientacoes_proposicao_camara}
\alias{fetch_orientacoes_proposicao_camara}
\title{Importa e processa dados das orientações das bancadas na Câmara dos Deputados para uma proposição}
\usage{
fetch_orientacoes_proposicao_camara(id_proposicao)
}
\arguments{
\item{id_proposicao}{Id da prposição para obtenção das orientações}
}
\value{
Dataframe contendo id da votação, id e voto dos deputados que participaram de cada votação
}
\description{
A partir do ID da proposição recupera votações que ocorreram no ano passado como parâmetro
}
\examples{
orientacoes <- fetch_orientacoes_por_proposicao_camara(2265603)
}
