% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_proposicoes_apresentadas_camara.R
\name{fetch_proposicoes_apresentadas_camara}
\alias{fetch_proposicoes_apresentadas_camara}
\title{Proposições de um tema que foram apresentadas na Câmara}
\usage{
fetch_proposicoes_apresentadas_camara(
  data_inicio = "2019-02-01",
  data_final = "2019-12-31",
  tema = "48"
)
}
\arguments{
\item{data_inicio}{Data de início para apresentação das proposições (formato yyyy-mm-dd)}

\item{data_final}{Data de fim para apresentação das proposições (formato yyyy-mm-dd)}

\item{tema}{Código do tema das proposições. Lista disponível em: https://dadosabertos.camara.leg.br/api/v2/referencias/proposicoes/codTema}
}
\value{
Dataframe com informações das proposições
}
\description{
Recupera lista de proposições de um tema que foram apresentadas a partir de uma determinada data 
na Câmara dos deputados.
}
\examples{
fetch_proposicoes_apresentadas_camara(data_inicio = "2019-02-01", data_final = "2019-12-31", tema = "48")
}
