% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_proposicoes_tramitadas_camara.R
\name{fetch_proposicoes_tramitadas_camara}
\alias{fetch_proposicoes_tramitadas_camara}
\title{Proposições de um tema que tramitaram na Câmara}
\usage{
fetch_proposicoes_tramitadas_camara(
  data_inicio = "2019-02-01",
  data_final = "2019-12-31",
  tema = "48",
  apenas_apresentada = FALSE
)
}
\arguments{
\item{data_inicio}{Data de início para ocorrência de eventos na tramitação das proposições (formato yyyy-mm-dd)}

\item{data_final}{Data de fim para ocorrência de eventos na tramitação das proposições (formato yyyy-mm-dd)}

\item{tema}{Código do tema das proposições. Lista disponível em: https://dadosabertos.camara.leg.br/api/v2/referencias/proposicoes/codTema}

\item{apenas_apresentada}{TRUE se o retorno deve considerar as proposições apresentadas no período passado como parâmetro.}
}
\value{
Dataframe com informações das proposições
}
\description{
Recupera lista de proposições de um tema que tramitaram a partir de uma determinada data 
na Câmara dos deputados.
}
\examples{
fetch_proposicoes_tramitadas_camara(data_inicio = "2019-02-01", data_final = "2019-12-31", tema = "48")
}
