% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_proposicoes_votadas_camara.R
\name{fetch_proposicoes_votadas_camara}
\alias{fetch_proposicoes_votadas_camara}
\title{Lista de proposições votadas em um determinado ano}
\usage{
fetch_proposicoes_votadas_camara(ano = 2019)
}
\arguments{
\item{ano}{Ano de ocorrência das votações}
}
\value{
Dataframe contendo id da proposição e data da última votação
}
\description{
Lista as proposições votadas em plenário para um determinado ano
}
\examples{
proposicoes_votadas_em_2019 <- fetch_proposicoes_votadas_camara(2019)
}
