% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_senadores_info.R
\name{fetch_senador}
\alias{fetch_senador}
\title{Importa dados particulares de um senador a partir da API de dados abertos do Senado}
\usage{
fetch_senador(id_senador)
}
\arguments{
\item{id_senador}{ID do senador na API de dados abertos do Senado}
}
\value{
Dataframe contendo informações do senador como nome e data de nascimento
}
\description{
Captura informações individuais dos senadores disponíveis apenas no endpoint específico para o Senador
}
\examples{
contarato <- fetch_senador(5953)
}
