% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_senadores_info.R
\name{fetch_senadores_atuais}
\alias{fetch_senadores_atuais}
\title{Importa dados de todos os senadores atualmente em exercício no Senado}
\usage{
fetch_senadores_atuais(legislatura_atual = 56)
}
\arguments{
\item{legislatura_atual}{Número da legislatura atual}
}
\value{
Dataframe contendo informações dos senadores: id, nome eleitoral, legislatura atual
}
\description{
Importa os dados de todos os senadores atualmente em exercício no Senado
}
\examples{
senadores <- fetch_senadores_atuais()
}
