% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_senadores_info.R
\name{fetch_senadores_legislatura}
\alias{fetch_senadores_legislatura}
\title{Importa dados de todos os senadores de uma legislatura específica}
\usage{
fetch_senadores_legislatura(legislatura = 56)
}
\arguments{
\item{legislatura}{Legislatura para recuperação da lista de senadores}
}
\value{
Dataframe contendo informações dos senadores: id, nome civil, uf, partido, genero, dentre outras
}
\description{
Importa os dados de todos os senadores de uma legislatura específica
}
\examples{
senadores <- fetch_senadores_legislatura(56)
}
