% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_votacoes_camara.R
\name{fetch_votacao}
\alias{fetch_votacao}
\title{Recupera informações de votações a partir de um id de votação}
\usage{
fetch_votacao(id_votacao)
}
\arguments{
\item{id_votacao}{ID da votação}
}
\value{
Info sobre uma votação
}
\description{
A partir do id de uma votação,
recupera seus dados disponíveis
}
\examples{
votacao_2190237_117 <- fetch_votacao("2190237-117")
}
