% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_votos_senado.R
\name{fetch_votos_por_proposicao_senado}
\alias{fetch_votos_por_proposicao_senado}
\title{Extrai informações de votos dos senadores a partir de um conjutno de votações}
\usage{
fetch_votos_por_proposicao_senado(id_proposicao)
}
\arguments{
\item{votacoes_senado_filepath}{Caminho para o csv das votações}
}
\value{
Dataframe com informações de votos dos senadores
}
\description{
A partir de um dataframe de votações, extrai os dados de votos dos senadores.
}
\examples{
votos <- fetch_votos_por_proposicao_senado(136635)
}
