% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_votos_senado.R
\name{fetch_votos_por_proposicao_votacao_senado}
\alias{fetch_votos_por_proposicao_votacao_senado}
\title{Extrai informações de votos dos senadores a partir dos ids de proposicao e de votacao}
\usage{
fetch_votos_por_proposicao_votacao_senado(id_proposicao, id_votacao)
}
\arguments{
\item{id_proposicao}{id da proposicao requerida}

\item{id_votacao}{id da votacao requerida}
}
\value{
Dataframe com informações de votos dos senadores
}
\description{
A partir de uma url base, extrai os dados de votos dos senadores por meia de seus ids
}
