% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_votos_camara.R
\name{fetch_votos_por_votacao_camara}
\alias{fetch_votos_por_votacao_camara}
\title{Importa e processa dados de votações na Câmara dos Deputados}
\usage{
fetch_votos_por_votacao_camara(id_votacao)
}
\arguments{
\item{id_votacao}{Id da votação no contexto interno do Voz Ativa}
}
\value{
Dataframe contendo id da votação, id e voto e partido dos deputados que participaram de cada votação
}
\description{
Recebe informações da votação específica para obtenção dos votos
}
\examples{
id_votacao = "2265603-43"
votacoes <- fetch_votos_por_votacao_camara("2179189-88")
}
