% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_proposicoes_votadas_camara.R
\name{fetch_xml_api_votacao_camara}
\alias{fetch_xml_api_votacao_camara}
\title{Recupera xml com votações de uma proposição específica}
\usage{
fetch_xml_api_votacao_camara(id_proposicao)
}
\arguments{
\item{id_proposicao}{ID da proposição}
}
\value{
XML com as votações da proposição
}
\description{
A partir do id da prosição recupera o xml com as votações em plenário da proposição
}
\examples{
votacoes_mpv8712019 <- fetch_xml_api_votacao_camara(2190355)
}
