% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_proposicoes_votadas_senado.R
\name{fetcher_votacoes_por_intervalo_senado}
\alias{fetcher_votacoes_por_intervalo_senado}
\title{Recupera informações das votações nominais do plenário em um intervalo de tempo}
\usage{
fetcher_votacoes_por_intervalo_senado(
  initial_date = "01/02/2019",
  end_date = format(Sys.Date(), "\%d/\%m/\%Y")
)
}
\arguments{
\item{initial_date}{Data inicial do período de votações. Formato: "dd/mm/yyyy"}

\item{end_date}{Data final do período de votações. Formato: "dd/mm/yyyy"}
}
\value{
Votações da proposição em um intervalo de tempo
}
\description{
A partir de uma data de início e uma de fim, recupera dados de
votações nominais de plenário que aconteceram na Senado Federal
}
\examples{
votacoes <- fetcher_votacoes_por_intervalo_senado()
}
