% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_proposicoes_votadas_senado.R
\name{fetcher_votacoes_por_proposicao_senado}
\alias{fetcher_votacoes_por_proposicao_senado}
\title{Recupera informações das votações nominais do plenário de uma proposição}
\usage{
fetcher_votacoes_por_proposicao_senado(id_proposicao, ano = NULL)
}
\arguments{
\item{id_proposicao}{Id da proposição}

\item{ano}{Ano que filtra as proposições. Caso seja null, todas as votações serão retornadas}
}
\value{
Votações de uma proposição em um ano ou todos, caso nenhum ano seja passado como parâmetro
}
\description{
A partir do id de uma proposição, recupera dados de
votações nominais de plenário que aconteceram na Senado Federal
}
\examples{
votacoes <- fetcher_votacoes_por_proposicao_senado(id_proposicao = 135251)
}
