% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_disciplina.R
\name{get_parlamentares_info}
\alias{get_parlamentares_info}
\title{Get parlamentares info}
\usage{
get_parlamentares_info()
}
\value{
Dataframe com dados dos parlamentares
}
\description{
Processa os parlamentares e filtra seus dados
em sua ultima legislatura
}
\examples{
get_parlamentares_info()
}
