% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sigla_by_id_camara}
\alias{get_sigla_by_id_camara}
\title{Recupera o título de uma proposição a partir de seu id}
\usage{
get_sigla_by_id_camara(id_proposicao)
}
\arguments{
\item{id_proposicao}{Id da proposição}
}
\value{
Título da proposicao (ex: MPV 867/2018)
}
\description{
Recebe o id de uma proposição na câmara e retorna seu título
}
\examples{
get_sigla_by_id_camara(2190237) // "MPV 867/2018"
}
