% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_orientacoes.R
\name{get_voto_lider}
\alias{get_voto_lider}
\title{Define a orientação de um partido para um conjunto de votos}
\usage{
get_voto_lider(lideres, votos, id_votacao)
}
\arguments{
\item{votos}{Dataframe contendo dados de votos}

\item{sigla_partido}{Sigla do partido a ter orientações recuperadas}
}
\value{
Dataframe contendo informações de orientações de um partido
}
\description{
Recupera informação das orientações dos partidos para um conjunto de votos utilizando a maioria absoluta.
No caso de empate entre os votos, o voto será o do líder do partido.
}
