% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapeia_nome_eleitoral_to_id_senado}
\alias{mapeia_nome_eleitoral_to_id_senado}
\title{Mapeia um nome eleitoral para id correspondente}
\usage{
mapeia_nome_eleitoral_to_id_senado(senadores_df, target_df)
}
\arguments{
\item{senadores_df}{Dataframe com as informações de nome_eleitoral e id}

\item{target_df}{Dataframe a receber o id do parlamentar}
}
\value{
Dataframe target_df contendo coluna id
}
\description{
Recebe dois dataframes contendo nome eleitoral e um deles com informação de id
}
