% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_sigla_bloco}
\alias{mutate_sigla_bloco}
\title{Separa orientações de bancadas compostas}
\usage{
mutate_sigla_bloco(orientacoes)
}
\arguments{
\item{orientacoes}{Dataframe de orientações, contendo a coluna partido_bloco}
}
\value{
Dataframe contendo a sigla de cada partido
}
\description{
Caso uma bancada possua mais de um partido, esta função retorna um dataset contendo as informações de cada bancada
separadamente
}
\examples{
mutate_sigla_bloco(orientacoes)
}
