% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{padroniza_nome}
\alias{padroniza_nome}
\title{Padroniza os nomes, retirando acentos, cedilhas e colcoando todas as letras em uppercase}
\usage{
padroniza_nome(nome)
}
\arguments{
\item{nome}{Nome a ser padronizado}
}
\value{
Nome padronizado
}
\description{
Recebe um nome e o padroniza no formato: sem acentos, cedilhas, letras maiúsculas
}
\examples{
padroniza_nome("çíço do álcórdéón")
}
