% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{padroniza_sigla}
\alias{padroniza_sigla}
\title{Padroniza siglas de partidos}
\usage{
padroniza_sigla(sigla)
}
\arguments{
\item{sigla}{Sigla do partido}
}
\value{
Dataframe com sigla do partido padronizada
}
\description{
Recebe uma sigla de partido como input e retorna seu valor padronizado
}
