% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{padroniza_texto}
\alias{padroniza_texto}
\title{Padroniza texto, retirando links, menções, pontuações, retirando acentos, números,
cedilhas e colocando todas as letras em lowercase}
\usage{
padroniza_texto(texto)
}
\arguments{
\item{texto}{Texto a ser padronizado}
}
\value{
Texto padronizado
}
\description{
Recebe um texto e o padroniza no formato: sem acentos, cedilhas, letras maiúsculas, links, números e menções
}
\examples{
padroniza_texto("çíço do álcórdéón")
}
