% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_disciplina.R
\name{participou_votacoes}
\alias{participou_votacoes}
\title{Participou votações}
\usage{
participou_votacoes(votos, orientacoes, enumera_orientacao)
}
\arguments{
\item{votos}{Dataframe de votos.}

\item{orientacoes}{Dataframe de orientações}

\item{enumera_orientacao}{Flag indicando se as orientações 
precisam ser transformadas em enum.}
}
\value{
Dataframe com id e casa do parlamentar e quantas votações ele participou
}
\description{
Analisa se o parlamentar participou em mais de 10 votações
}
\examples{
participou_votacoes(votos, orientacoes, enumera_orientacao)
}
