% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_orientacoes.R
\name{process_orientacao_anos_camara}
\alias{process_orientacao_anos_camara}
\title{Processa orientação de proposições (selecionadas pela equipe VA) votadas em plenário para um conjunto de anos}
\usage{
process_orientacao_anos_camara(anos = c(2019, 2020, 2021))
}
\arguments{
\item{anos}{Vector com lista de anos}
}
\value{
Lista contendo orientações
}
\description{
Recupera informação das orientações dos partidos para um conjunto de anos para as
proposicoes selecionadas pela equipe VA.
}
\examples{
orientacao <- process_orientacao_anos_url_camara(2019)
}
