% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_orientacoes.R
\name{process_orientacao_por_ano_camara}
\alias{process_orientacao_por_ano_camara}
\title{Processa orientação de proposições votadas em plenário para um determinado ano}
\usage{
process_orientacao_por_ano_camara(ano = 2019)
}
\arguments{
\item{ano}{Ano para ocorrência das votações em plenário}

\item{url}{Link para dados das proposições selecionadas para captura das votações em plenário
Se url é diferente de NULL, então considerará lista de proposições presentes nos dados disponíveis através da URL}
}
\value{
Dataframe com informações das orientações
}
\description{
Recupera informação das orientações dos partidos para um determinado ano
}
\examples{
orientacao <- process_orientacao_por_ano_camara(2019)
}
