% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_orientacoes.R
\name{process_orientacao_votos_senado}
\alias{process_orientacao_votos_senado}
\title{Processa orientação dos votos de plenário para um conjunto de votações}
\usage{
process_orientacao_votos_senado(votos)
}
\arguments{
\item{votos_datapath}{Caminho para o dataframe contendo dados de votações}
}
\value{
Lista contendo orientações
}
\description{
Recupera informação das orientações dos partidos para um conjunto de votações
}
\examples{
votos <- processa_votos_senado()
orientacao <- process_orientacao_votos_senado()
}
