% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_disciplina.R
\name{processa_bancada_suficiente}
\alias{processa_bancada_suficiente}
\title{Processa bancada suficiente}
\usage{
processa_bancada_suficiente(minimo_deputados = 5, minimo_senadores = 3)
}
\value{
Dataframe com partidos e se sua bancada é suficiente
}
\description{
Processa os parlamentares e filtra aqueles que não bancada
suficiente na camara/senado.
}
\examples{
processa_bancada_suficiente()
}
