% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_cargos_mesa.R
\name{processa_cargos_mesa}
\alias{processa_cargos_mesa}
\title{Captura informações de deputados e senadores atualmente 
integrantes da Mesa na Câmara de Deputados e no Senado Federal}
\usage{
processa_cargos_mesa(casa = NULL)
}
\value{
Dataframe contendo deputados senadores com cargos na Mesa Diretora da
Câmara e Senado.
}
\description{
Com base na API da Câmara e do Senado,
 captura informações dos integrantes da Mesa Diretora em cada casa.
}
\examples{
processa_cargos_mesa()
}
