% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_comissoes.R
\name{processa_comissoes_e_composicoes}
\alias{processa_comissoes_e_composicoes}
\title{Recupera informações das Comissões e de suas composições}
\usage{
processa_comissoes_e_composicoes(casa = NULL)
}
\value{
Lista com dois Dataframes: comissões e composição das comissões
}
\description{
Retorna dados de Comissões da Câmara dos Deputados e também suas composições
}
\examples{
processa_comissoes()
}
