% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_parlamentar.R
\name{processa_dados_deputados}
\alias{processa_dados_deputados}
\title{Processa dados de deputados}
\usage{
processa_dados_deputados()
}
\value{
Dataframe contendo informações sobre os deputados
}
\description{
Processa informações sobre os deputados das legislaturas 55 e 56
}
\examples{
#' processa_dados_deputados()
}
