% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_parlamentar.R
\name{processa_dados_parlamentares}
\alias{processa_dados_parlamentares}
\title{Processa dados de parlamentares}
\usage{
processa_dados_parlamentares()
}
\value{
Dataframe contendo informações sobre os parlamentares (deputados e senadores)
}
\description{
Processa informações sobre os parlamentares da legislatura atual
}
\examples{
processa_dados_parlamentares()
}
