% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_parlamentar.R
\name{processa_dados_senadores}
\alias{processa_dados_senadores}
\title{Processa dados de senadores}
\usage{
processa_dados_senadores()
}
\value{
Dataframe contendo informações sobre os senadores
}
\description{
Processa informações sobre os senadores das legislaturas 55 e 56
}
\examples{
processa_dados_senadores()
}
