% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_financiamento.R
\name{processa_financiamento_por_setores}
\alias{processa_financiamento_por_setores}
\title{Processa dados de financiamento dos parlamentares}
\usage{
processa_financiamento_por_setores(
  parlamentares_doadores,
  info_empresas,
  empresas_doadores
)
}
\arguments{
\item{parlamentares_doadores}{Dataframe com informação das doações recebidas por parlamentares
durante a campanha eleitoral. O csv de 2018 pode ser obtido aqui: https://github.com/parlametria/perfil-parlamentar-dados/blob/master/parlametria/raw_data/receitas/parlamentares_doadores.csv}

\item{info_empresas}{Dataframe com informação das empresas que tiveram sócios que doaram para
parlamentares na campanha eleitoral. O csv de 2018 pode ser obtido aqui: https://github.com/parlametria/perfil-parlamentar-dados/blob/master/parlametria/raw_data/empresas/info_empresas_doadores_todos_parlamentares.csv}

\item{empresas_doadores}{Dataframe com informação dos sócios de empresas que doaram para
parlamentares na campanha eleitoral. O csv de 2018 pode ser obtido aqui: https://github.com/parlametria/perfil-parlamentar-dados/blob/master/parlametria/raw_data/empresas/empresas_doadores_todos_parlamentares.csv}
}
\value{
Dataframe com doações feitas
}
\description{
Processa doações feitas por setores econômicos para parlamentares em 2018.
ATENÇÃO: Se uma mesma doação foi feita por um sócio de empresa ligada a mais de uma
atividade econômica, o valor da doação será dividido por igual para cada atividade.
}
\examples{
processa_financiamento_setores()
}
