% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_governismo.R
\name{processa_governismo}
\alias{processa_governismo}
\title{Calcula Governismo}
\usage{
processa_governismo(votos)
}
\arguments{
\item{votos}{Dataframe de votos para cálculo do Governismo. Os votos devem ser de uma mesma casa.
Ou seja ou câmara ou senado. Devem ter pelo menos 3 colunas: id_votacao, id_parlamentar e voto.}
}
\value{
Dataframe com parlamentares e as dimensões para o Governismo
}
\description{
Calcula Governismo usando a técnica de ideal points. 
Apenas parlamentares com 10 ou mais votos válidos tem seu governismo calculado.
}
\examples{
processa_governismo(votos)
}
