% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_liderancas.R
\name{processa_liderancas}
\alias{processa_liderancas}
\title{Processa dados de lideranças da Câmara e do Senado}
\usage{
processa_liderancas(casa = NULL)
}
\value{
Dataframe contendo informações das lideranças
}
\description{
Executa funções que recuperam dados de liderança na Câmara e no Senado.
}
\examples{
liderancas <- processa_liderancas()
}
