% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_mandatos.R
\name{processa_mandatos_parlamentar}
\alias{processa_mandatos_parlamentar}
\title{Extrai informações sobre o mandato de um parlamentar}
\usage{
processa_mandatos_parlamentar(id_parlamentar, casa)
}
\arguments{
\item{id_parlamentar}{id do parlamentar}

\item{casa}{casa do parlamentar}
}
\value{
Dataframe contendo informações de legislatura,
data de inicio, data de fim, situacao, código e descrição da causa do fim do exercício.
}
\description{
Recebe o id e a casa de um parlamentar e retorna um dataframe contendo informações do mandato, como legislatura,
data de inicio, data de fim, situacao, código e descrição da causa do fim do exercício.
}
\examples{
processa_mandatos_parlamentar(141428, "camara")
}
