% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_mandatos.R
\name{processa_mandatos_parlamentares}
\alias{processa_mandatos_parlamentares}
\title{Extrai informações sobre os mandato de todos os parlamentares}
\usage{
processa_mandatos_parlamentares(df_parlamentares = NULL)
}
\arguments{
\item{df_parlamentares}{dataframe com informações de id e casa dos parlamentares}
}
\value{
Dataframe contendo informações de legislatura,
data de inicio, data de fim, situacao, código e descrição da causa do fim do exercício.
}
\description{
Recebeum dataframe contendo id e casa dos parlamentares e retorna um dataframe contendo informações do mandato, como legislatura,
data de inicio, data de fim, situacao, código e descrição da causa do fim do exercício.
}
\examples{
processa_mandatos_parlamentares(perfilparlamentar::fetch_deputados_legislatura(56))
}
