% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_disciplina.R
\name{processa_num_votacoes_parlamentares}
\alias{processa_num_votacoes_parlamentares}
\title{Processa num votações parlamantares}
\usage{
processa_num_votacoes_parlamentares(votos, orientacoes, enumera_orientacao)
}
\arguments{
\item{votos}{Dataframe de votos
Os votos devem ter pelo menos 2 colunas: id_votacao e voto.}

\item{orientacoes}{Dataframe de orientações}

\item{enumera_orientacao}{Flag indicando se as orientações 
precisam ser transformadas em enum.}
}
\value{
Dataframe com o id do parlamentar e quantas vezes ele votou
}
\description{
Processa o número total de votações para cada parlamentar
}
\examples{
processa_num_votacoes_parlamentares(votos, orientacoes)
}
