% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_financiamento.R
\name{processa_socios_empresas_setores}
\alias{processa_socios_empresas_setores}
\title{Processa doações por setor}
\usage{
processa_socios_empresas_setores(info_empresas, empresas_doadores)
}
\arguments{
\item{info_empresas}{Dataframe com informação das empresas que tiveram sócios que doaram para
parlamentares na campanha eleitoral.}

\item{empresas_doadores}{Dataframe com informação dos sócios de empresas que doaram para
parlamentares na campanha eleitoral.}
}
\value{
Dataframe com doações feitas
}
\description{
Processa doações feitas por setores econômicos para parlamentares em 2018.
}
\examples{
processa_socios_empresas_setores()
}
