% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_votacoes.R
\name{processa_votacoes_camara_anos}
\alias{processa_votacoes_camara_anos}
\title{Processa todas as votações da Câmara para um conjunto de anos}
\usage{
processa_votacoes_camara_anos(anos = c(2019, 2020))
}
\arguments{
\item{anos}{Lista de anos para recuperação das votações}
}
\value{
Dataframe votações
}
\description{
A partir de uma lista de anos recupera as votações da câmara
}
\examples{
processa_votacoes_camara_anos(anos = c(2020))
}
