% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_disciplina.R
\name{processa_votacoes_sem_consenso}
\alias{processa_votacoes_sem_consenso}
\title{Processa votações sem consenso}
\usage{
processa_votacoes_sem_consenso(votos, limite_consenso = 0.9)
}
\arguments{
\item{votos}{Dataframe de votos. Os votos devem ser de uma mesma casa.
Ou seja ou câmara ou senado. Devem ter pelo menos 2 colunas: id_votacao e voto.}
}
\value{
Dataframe com votações e as proporções de consenso
}
\description{
Processa as votações para filtrar quais não há consenso
}
\examples{
processa_votacoes_sem_consenso(votos)
}
