% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_votacoes_sumarizadas.R
\name{processa_votacoes_sumarizadas}
\alias{processa_votacoes_sumarizadas}
\title{Processa votações sumarizadas}
\usage{
processa_votacoes_sumarizadas(votos)
}
\arguments{
\item{votos}{Dataframe de votos. Devem ter pelo menos 2 colunas: id_votacao e voto.}
}
\value{
Dataframe com ids de parlamentares e quantidade de votos
}
\description{
Processa o número de votações por parlamentar com filtros de governismo e disciplina
}
\examples{
processa_votacoes_sumarizadas(votos)
}
