% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_votos.R
\name{processa_votos_camara_anos}
\alias{processa_votos_camara_anos}
\title{Processa todos os votos da Câmara para um conjunto de anos}
\usage{
processa_votos_camara_anos(anos = c(2019, 2020))
}
\arguments{
\item{anos}{Lista de anos para recuperação dos votos}
}
\value{
Dataframe votos dos parlamentares
}
\description{
A partir de uma lista de anos recupera os votos da câmara
}
\examples{
processa_votos_camara_anos(anos = c(2020))
}
