% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_disciplina.R
\name{processa_votos_orientados}
\alias{processa_votos_orientados}
\title{Processa votos orientados}
\usage{
processa_votos_orientados(votos, orientacoes)
}
\arguments{
\item{votos}{Dataframe de votos
Os votos devem ter pelo menos 2 colunas: id_votacao e voto.}

\item{orientacoes}{Dataframe de orientações}
}
\value{
Dataframe com o que cada parlamentar votou e qual era a orientação do partido
}
\description{
Processa os votos e suas orientações
}
\examples{
processa_votos_orientados(votos, orientacoes)
}
