% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_votos.R
\name{processa_votos_senado}
\alias{processa_votos_senado}
\title{Recupera dados dos votos para um conjunto de votações.}
\usage{
processa_votos_senado(votos, senadores_df)
}
\arguments{
\item{votos}{Dataframe com os votos no Senado}

\item{senadores_df}{Dataframe dos senadores}
}
\value{
Datafrane com os votos.
}
\description{
Recebe um caminho para o dataframe de votações e retorna todos os
votos relacionados.
}
\examples{
votos <- fetch_votos_por_proposicao_senado(136635)
senadores_df <- read_csv("https://raw.githubusercontent.com/parlametria/relatorio-bianuario-ma/a5d55b0bda60b8a4f6f8317adc9e29148b1ab0a0/data/raw/leggo_data/entidades.csv")
senadores_df <- (senadores_df \%>\% 
      filter(casa == 'senado', is_parlamentar == 1) \%>\% 
      select(id_parlamentar = id_entidade, nome_eleitoral = nome))
votos_com_id_senador <- process_votos_senado(votos, senadores_df)
}
