% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_votos.R
\name{processa_votos_senado_anos}
\alias{processa_votos_senado_anos}
\title{Processa todos os votos do Senado para um conjunto de anos}
\usage{
processa_votos_senado_anos(anos = c(2019, 2020))
}
\arguments{
\item{anos}{Lista de anos para recuperação dos votos}
}
\value{
Dataframe votos dos parlamentares
}
\description{
A partir de uma lista de anos recupera os votos do Senado
}
\examples{
processa_votos_senado_anos(anos = c(2020))
}
