% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader_financiamento.R
\name{read_parlamentares_doadores}
\alias{read_parlamentares_doadores}
\title{Lê csv de parlamentares doadores}
\usage{
read_parlamentares_doadores(parlamentares_doadores_path)
}
\arguments{
\item{parlamentares_doadores_path}{Caminho para o CSV com informação das doações recebidas por parlamentares
durante a campanha eleitoral. 
O csv de 2018 pode ser obtido aqui: 
https://raw.githubusercontent.com/parlametria/perfil-parlamentar-dados/master/parlametria/raw_data/receitas/parlamentares_doadores.csv}
}
\value{

}
\description{
Lê csv de parlamentares doadores
}
\examples{
read_parlamentares_doadores("https://raw.githubusercontent.com/parlametria/perfil-parlamentar-dados/master/parlametria/raw_data/receitas/parlamentares_doadores.csv")
}
