% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_empty_list_for_na}
\alias{replace_empty_list_for_na}
\title{Retorna um NA caso o campo seja uma lista nula}
\usage{
replace_empty_list_for_na(valor)
}
\arguments{
\item{valor}{valor a ser checado}
}
\description{
Recebe um valor e retorna NA se for uma lista nula ou
o próprio valor, caso contrário.
}
