% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_comissoes_senado.R
\name{fetch_comissoes_composicao_senado}
\alias{fetch_comissoes_composicao_senado}
\title{Recupera informações de composição das comissões usando o LeggoR}
\usage{
fetch_comissoes_composicao_senado()
}
\value{
Dataframe com parlamentares membros da comissão e seus respectivos cargos
}
\description{
Utiliza o LeggoR para recuperar informações sobre a composição dos membros das composições do Congresso
}
\examples{
fetch_comissoes_composicao_senado()
}
