% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_proposicoes_votadas_camara.R
\name{fetch_votacoes_por_proposicao_camara}
\alias{fetch_votacoes_por_proposicao_camara}
\title{Recupera informações de votações em plenário a partir de um id de proposição}
\usage{
fetch_votacoes_por_proposicao_camara(id_proposicao)
}
\arguments{
\item{id_proposicao}{ID da proposição}
}
\value{
Info sobre as votações
}
\description{
A partir do id de uma proposição,
recupera dados de todas as votações disponíveis
}
\examples{
votacoes_mpv8712019 <- fetch_votacoes_por_proposicao_camara(2190355)
}
