% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher_proposicoes_apresentadas_senado.R
\name{fetcher_proposicoes_em_intervalo_senado}
\alias{fetcher_proposicoes_em_intervalo_senado}
\title{Recupera dados das proposições que foram apresentadas em
um intervalo de datas.}
\usage{
fetcher_proposicoes_em_intervalo_senado(
  data_inicial = "20200311",
  data_final = gsub("-", "", Sys.Date())
)
}
\arguments{
\item{data_inicial}{Data inicial do intervalo (formato AAAAMMDD)}

\item{data_final}{Data final do intervalo (formato AAAAMMDD)}
}
\value{
Dataframe com lista de proposições apresentadas em um intervalo
de datas.
}
\description{
Recebe uma data inicial e uma data final, retorna as 
proposições apresentadas no intervalo.
}
