% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_disciplina.R
\name{participou_votacoes}
\alias{participou_votacoes}
\title{Participou votações}
\usage{
participou_votacoes(parlamentar_id, votos)
}
\arguments{
\item{parlamentar_id}{Id de um parlamentar.}

\item{votos}{Dataframe de votos.}
}
\value{
TRUE caso o parlamentar tenha participado em mais de 10 votações
}
\description{
Analisa se o parlamentar participou em mais de 10 votações
}
\examples{
participou_votacoes(285, votos)
}
