% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlamentares_por_legislatura.R
\name{processa_dados_parlamentares}
\alias{processa_dados_parlamentares}
\title{Processa dados de parlamentares}
\usage{
processa_dados_parlamentares(casa = NULL, legislaturas_list = c(55, 56))
}
\arguments{
\item{casa}{Casa (câmara ou senado) para retornar os dados. Se nenhuma for passada,
a função retornará dados de ambas.}

\item{legislaturas_list}{Lista com ids das legislaturas. Default = 55 e 56.}
}
\value{
Dataframe contendo informações sobre os parlamentares (deputados e senadores)
}
\description{
Processa informações sobre os parlamentares da legislatura atual
}
\examples{
processa_dados_parlamentares("camara", c(56))
processa_dados_parlamentares("senado", c(55))
}
