% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer_disciplina.R
\name{processa_disciplina_partidaria}
\alias{processa_disciplina_partidaria}
\title{Processa disciplina partidária}
\usage{
processa_disciplina_partidaria(votos, orientacoes, enumera_orientacao)
}
\arguments{
\item{votos}{Dataframe de votos
Os votos devem ter pelo menos 2 colunas: id_votacao e voto.}

\item{orientacoes}{Dataframe de orientações}

\item{enumera_orientacao}{Flag indicando se as orientações 
precisam ser transformadas em enum.}
}
\value{
Dataframe de parlamentares e sua disciplina partidária
}
\description{
Processa as votações para filtrar quais não há consenso
}
\examples{
processa_disciplina_partidaria(votos, orientacoes, enumera_orientacao)
}
